+++ CHANGELOG.md
@@ -11,6 +11,8 @@ incremented for features.
 
 ## [Unreleased]
 
+## [Unreleased Breaking]
+
 * spl: Remove the `shmem` program from the `spl` crate ([965](https://github.com/project-serum/anchor/pull/965))
 * lang: Remove program `AccountInfo` from accounts list for CPI because it is no longer needed ([965](https://github.com/project-serum/anchor/pull/965))